gdb: fix typo in NEWS file
The following commit introduced a typo to the NEW file:

  commit d4f644c8416f8d60942f9e650fa21a473eb8e06f
  Date:   Sat Mar 15 12:03:50 2025 +0000

      gdb/python: remove unused argument from builtin_disassemble

this commit fixes it.

I've also reworded the NEWS entry a little.  Simon pointed out in
review that the unused argument was also documented in Python's help()
output, which I hadn't mentioned in the NEWS entry.  I've updated the
NEWS entry to just highlight that the now removed argument was never
mentioned in the manual, I think that's all that really matters.