Remove cooked_index_worker::result_type
cooked_index_worker::result_type is an ad hoc tuple type used for
transferring data between phases of the indexer.  It's a bit unwieldy
and another patch I'm working on would be somewhat nicer without it.

This patch removes the type.  Now cooked_index_ephemeral objects are
transferred instead, which is handy because they already hold the
needed state.

Approved-By: Simon Marchi <simon.marchi@efficios.com>