gdb/dwarf: remove redundant read of dwo_name
lookup_dwo_unit receives the name of the DWO unit to look up, as read
from the DW_AT_dwo_name attribute of the skeleton DIE.  But then, it
doesn't use it:

    /* Yeah, we look dwo_name up again, but it simplifies the code.  */
    dwo_name = dwarf2_dwo_name (comp_unit_die, cu);

Perhaps this comment made sense at some point, but with the code we have
today, I don't understand it.  It should be fine to use the name passed
as a parameter, which the caller also obtained by calling
dwarf2_dwo_name.

Change-Id: I84723e12726f77e4202d042428ee0eed9962ceb8
Approved-By: Tom Tromey <tom@tromey.com>