gdb/dwarf: use reference in cutu_reader::cutu_reader interface
Change some parameters to be references instead of pointers, when the
value must not be nullptr.  I'd like to do this more of this kind of
change, but I have to limit the scope of the change, otherwise there's
just no end (and some local variables could also be turned into
references).  So for now, just do it the cutu_reader constructors.

Change-Id: I9442c6043726981d58f9b141f516c590c0a71bcc
Approved-By: Tom Tromey <tom@tromey.com>