Make gdb/guile codespell-clean
This cleans up the last codespell reports in the Guile directory and
adds gdb/guile to pre-commit.

It also tells codespell to ignore URLs.  I think this is warranted
because many URLs don't really contain words per se; and furthermore
if any URL-checking is needed at all, it would be for liveness and not
spelling.

Also I was wondering why the codespell config is in contrib and not
gdb/setup.cfg.

Approved-By: Tom de Vries <tdevries@suse.de>